+++ src/util.ts
@@ -1,2 +1,3 @@
 export { figure } from './util/figure';
+export { footnote } from './util/footnote';
 export { toc } from './util/toc';